Move interface types to shared folder
