--- src/renderer/ConfigTemplate.d.ts
@@ -1,46 +1,0 @@
-import { IClientAuthOptions } from 'libsixgrid/dist/src/Client'
-import { PostRating } from 'libsixgrid'
-import { MetricManagerData } from './MetricManager'
-import { KeybindProfile } from './Keybinder/KeybindProfile'
-
-export interface AuthTemplate
-{
-    auth: IClientAuthOptions
-    endpoint: string
-}
-export interface IConfig_AuthProfile
-{
-    items: AuthTemplate[]
-    _current: number
-}
-
-export interface IConfig_User
-{
-    media: {
-        autoplay: boolean,
-        loop: boolean
-    },
-    downloadFolder: string
-    saveMetadata: boolean
-    tagBlacklist: string[]
-    ratingFilter: PostRating
-    preloadPageCount: number
-    preloadStartIndex: number
-    highQualityPreview: boolean
-    sortByScore: boolean
-    sortByFavorite: boolean
-    ratingSafe: boolean
-    ratingQuestionable: boolean
-    ratingExplicit: boolean
-    zoomFactor: number // float
-}
-
-export interface IConfig_Stats
-{
-    metricStore: MetricManagerData
-}
-export interface IConfig_Keybind
-{
-    currentProfile: string
-    currentProfileData: KeybindProfile
-}
