+++ src/renderer/Keybinder/KeybindProfile.ts
@@ -1,11 +1,8 @@
-import {KeybindItem} from './KeybindItem'
+import { IKeybindProfile } from '../../shared/keybind'
+import { KeybindItem } from './KeybindItem'
+
 
 const toolbox = require('tinytoolbox')
-export interface IKeybindProfile
-{
-    Id: string
-    Name: string
-}
 export class KeybindProfile implements IKeybindProfile
 {
     Id: string = toolbox.stringGen(8)
