--- src/renderer/Keybinder/index.ts
@@ -1,14 +1,0 @@
-export declare type KeybindChord = number[]
-export declare type KeybindChannel =
-    'unknown' |
-    'item:next' |
-    'item:previous' |
-    'view:close'
-
-
-
-export enum Keystate
-{
-    Up,
-    Down
-}
