+++ src/renderer/index.d.ts
@@ -2,7 +2,7 @@ import Configuration, { IConfiguration } from './Configuration'
 import type {EventEmitter} from 'events'
 import Post from 'libsixgrid/dist/src/Post'
 import {IProductInformation} from '../shared'
-import { IConfig_AuthProfile, IConfig_Stats, IConfig_User, IConfig_Keybind } from './ConfigTemplate'
+import { IConfig_AuthProfile, IConfig_Stats, IConfig_User, IConfig_Keybind } from '../shared/config'
 import MetricManager from './MetricManager'
 import Steamworks from './SteamworksIntergration'
 import { KeybindManager } from './Keybinder/KeybindManager'
