+++ src/shared/metricManager.ts
@@ -1,0 +1,2 @@
+import { ISteamMetric } from "./steamworks"
+export type MetricManagerData = {[key: string]: ISteamMetric}
